Moved all 'parser'-functions to separate file
